utils/requirements_dev.txt
annotated-types==0.7.0
argcomplete==3.6.0
black==25.1.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
datamodel-code-generator==0.28.2
dicttoxml==1.7.16
genson==1.3.0
inflect==5.6.2
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.11.1
pydantic_core==2.33.0
PyYAML==6.0.2
typing_extensions==4.13.0
